third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-attribute-basic.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/popup-utils.js"></script>
<script src="../../resources/common.js"></script>

<div id=popups>
  <div popup id=boolean>Pop up</div>
  <div popup="">Pop up</div>
  <div popup=auto>Pop up</div>
  <div popup=hint>Pop up</div>
  <div popup=manual>Pop up</div>
  <article popup>Different element type</article>
  <header popup>Different element type</header>
  <nav popup>Different element type</nav>
  <input type=text popup value="Different element type">
  <dialog popup>Dialog with popup attribute</dialog>
  <dialog popup="manual">Dialog with popup=manual</dialog>
  <div popup=true>Invalid popup value - defaults to popup=manual</div>
  <div popup=popup>Invalid popup value - defaults to popup=manual</div>
  <div popup=invalid>Invalid popup value - defaults to popup=manual</div>
</div>

<div id=nonpopups>
  <div>Not a pop-up</div>
  <dialog open>Dialog without popup attribute</dialog>
</div>

<div popup class=animated>Animated pop-up</div>
<style>
[popup].animated {
  opacity: 0;
  transition: opacity 10s;
}
[popup].animated:open {
  opacity: 1;
}
</style>

<script>
  function assertPopUpVisibility(popUp, isPopUp, expectedVisibility, message) {
    const isVisible = isElementVisible(popUp);
    assert_equals(isVisible, expectedVisibility,`${message}: Expected this element to be ${expectedVisibility ? "visible" : "not visible"}`);
    // Check other things related to being visible or not:
    if (isVisible) {
      assert_not_equals(window.getComputedStyle(popUp).display,'none');
      assert_equals(popUp.matches(':open'),isPopUp,`${message}: Visible pop-ups should match :open`);
    } else {
      assert_equals(window.getComputedStyle(popUp).display,'none',`${message}: Non-showing pop-ups should have display:none`);
      assert_false(popUp.matches(':open'),`${message}: Non-showing pop-ups should *not* match :open`);
    }
  }
  function assertIsFunctionalPopUp(popUp) {
    assertPopUpVisibility(popUp, /*isPopUp*/true, /*expectedVisibility*/false, 'A pop-up should start out hidden');
    popUp.showPopUp();
    assertPopUpVisibility(popUp, /*isPopUp*/true, /*expectedVisibility*/true, 'After showPopUp(), a pop-up should be visible');
    assert_throws_dom("InvalidStateError",() => popUp.showPopUp(),'Calling showPopUp on a showing pop-up should throw InvalidStateError');
    popUp.hidePopUp();
    assertPopUpVisibility(popUp, /*isPopUp*/true, /*expectedVisibility*/false, 'After hidePopUp(), a pop-up should be hidden');
    assert_throws_dom("InvalidStateError",() => popUp.hidePopUp(),'Calling hidePopUp on a hidden pop-up should throw InvalidStateError');
    const parent = popUp.parentElement;
    popUp.remove();
    assert_throws_dom("InvalidStateError",() => popUp.showPopUp(),'Calling showPopUp on a disconnected pop-up should throw InvalidStateError');
    parent.appendChild(popUp);
  }
  function assertNotAPopUp(nonPopUp) {
    // If the non-pop-up element nonetheless has a 'popup' attribute, it should
    // be invisible. Otherwise, it should be visible.
    const expectVisible = !nonPopUp.hasAttribute('popup');
    assertPopUpVisibility(nonPopUp, /*isPopUp*/false, expectVisible, 'A non-pop-up should start out visible');
    assert_throws_dom("NotSupportedError",() => nonPopUp.showPopUp(),'Calling showPopUp on a non-pop-up should throw NotSupportedError');
    assertPopUpVisibility(nonPopUp, /*isPopUp*/false, expectVisible, 'Calling showPopUp on a non-pop-up should leave it visible');
    assert_throws_dom("NotSupportedError",() => nonPopUp.hidePopUp(),'Calling hidePopUp on a non-pop-up should throw NotSupportedError');
    assertPopUpVisibility(nonPopUp, /*isPopUp*/false, expectVisible, 'Calling hidePopUp on a non-pop-up should leave it visible');
  }

  // Start with the provided examples:
  Array.from(document.getElementById('popups').children).forEach(popUp => {
    test((t) => {
      assertIsFunctionalPopUp(popUp);
    }, `The element ${popUp.outerHTML} should behave as a pop-up.`);
  });
  Array.from(document.getElementById('nonpopups').children).forEach(nonPopUp => {
    test((t) => {
      assertNotAPopUp(nonPopUp);
    }, `The element ${nonPopUp.outerHTML} should *not* behave as a pop-up.`);
  });

  // Then loop through all HTML5 elements that render a box by default:
  let elementsThatDontRender = ['audio','base','br','datalist','dialog','embed','head','link','meta','noscript','param','rp','script','style','template','title','wbr'];
  const elements = HTML5_ELEMENTS.filter(el => !elementsThatDontRender.includes(el));
  elements.forEach(tag => {
    test((t) => {
      const element = document.createElement(tag);
      element.setAttribute('popup','auto');
      document.body.appendChild(element);
      t.add_cleanup(() => element.remove());
      assertIsFunctionalPopUp(element);
    }, `A <${tag}> element should behave as a pop-up.`);
  });

  function createPopUp(t) {
    const popUp = document.createElement('div');
    document.body.appendChild(popUp);
    t.add_cleanup(() => popUp.remove());
    popUp.setAttribute('popup','auto');
    return popUp;
  }

  test((t) => {
    // You can set the `popup` attribute to anything.
    // Setting the `popUp` IDL to a string sets the content attribute to exactly that, always.
    // Getting the `popUp` IDL value only retrieves valid values.
    const popUp = createPopUp(t);
    assert_equals(popUp.popUp,'auto');
    popUp.setAttribute('popup','hint');
    assert_equals(popUp.popUp,'hint');
    popUp.setAttribute('popup','HiNt');
    assert_equals(popUp.popUp,'hint','Case is normalized in IDL');
    assert_equals(popUp.getAttribute('popup'),'HiNt','Case is *not* normalized/changed in the content attribute');
    popUp.popUp='hInT';
    assert_equals(popUp.popUp,'hint','Case is normalized in IDL');
    assert_equals(popUp.getAttribute('popup'),'hInT','Value set from IDL is propagated exactly to the content attribute');
    popUp.setAttribute('popup','invalid');
    assert_equals(popUp.popUp,'manual','Invalid values should reflect as "manual"');
    popUp.removeAttribute('popup');
    assert_equals(popUp.popUp,null,'No value should reflect as null');
    popUp.popUp='hint';
    assert_equals(popUp.getAttribute('popup'),'hint');
    popUp.popUp='auto';
    assert_equals(popUp.getAttribute('popup'),'auto');
    popUp.popUp='';
    assert_equals(popUp.getAttribute('popup'),'');
    assert_equals(popUp.popUp,'auto');
    popUp.popUp='AuTo';
    assert_equals(popUp.getAttribute('popup'),'AuTo');
    assert_equals(popUp.popUp,'auto');
    popUp.popUp='invalid';
    assert_equals(popUp.getAttribute('popup'),'invalid','IDL setter allows any value');
    assert_equals(popUp.popUp,'manual','but IDL getter reflects "manual"');
    popUp.popUp='';
    assert_equals(popUp.getAttribute('popup'),'','IDL setter propagates exactly');
    assert_equals(popUp.popUp,'auto','Empty should map to auto in IDL');
    popUp.popUp='auto';
    popUp.popUp=null;
    assert_equals(popUp.getAttribute('popup'),null,'Setting null for the IDL property should remove the content attribute');
    assert_equals(popUp.popUp,null,'Null returns null');
    popUp.popUp='auto';
    popUp.popUp=undefined;
    assert_equals(popUp.getAttribute('popup'),null,'Setting undefined for the IDL property should remove the content attribute');
    assert_equals(popUp.popUp,null,'undefined returns null');
  },'IDL attribute reflection');

  test((t) => {
    const popUp = createPopUp(t);
    assertIsFunctionalPopUp(popUp);
    popUp.removeAttribute('popup');
    assertNotAPopUp(popUp);
    popUp.setAttribute('popup','AuTo');
    assertIsFunctionalPopUp(popUp);
    popUp.removeAttribute('popup');
    popUp.setAttribute('PoPuP','AuTo');
    assertIsFunctionalPopUp(popUp);
    // Via IDL also
    popUp.popUp = 'auto';
    assertIsFunctionalPopUp(popUp);
    popUp.popUp = 'aUtO';
    assertIsFunctionalPopUp(popUp);
    popUp.popUp = 'invalid'; // treated as "manual"
    assertIsFunctionalPopUp(popUp);
  },'Popup attribute value should be case insensitive');

  test((t) => {
    const popUp = createPopUp(t);
    assertIsFunctionalPopUp(popUp);
    popUp.setAttribute('popup','hint'); // Change pop-up type
    assertIsFunctionalPopUp(popUp);
    popUp.setAttribute('popup','invalid'); // Change pop-up type to something invalid
    assertIsFunctionalPopUp(popUp);
    popUp.popUp = 'hint'; // Change pop-up type via IDL
    assertIsFunctionalPopUp(popUp);
    popUp.popUp = 'invalid'; // Make invalid via IDL (treated as "manual")
    assertIsFunctionalPopUp(popUp);
  },'Changing attribute values for pop-up should work');

  test((t) => {
    const popUp = createPopUp(t);
    popUp.showPopUp();
    assert_true(popUp.matches(':open'));
    popUp.setAttribute('popup','hint'); // Change pop-up type
    assert_false(popUp.matches(':open'));
    popUp.showPopUp();
    assert_true(popUp.matches(':open'));
    popUp.setAttribute('popup','manual');
    assert_false(popUp.matches(':open'));
    popUp.showPopUp();
    assert_true(popUp.matches(':open'));
    popUp.setAttribute('popup','invalid');
    assert_true(popUp.matches(':open'),'From "manual" to "invalid" (which is interpreted as "manual") should not close the pop-up');
    popUp.setAttribute('popup','auto');
    assert_false(popUp.matches(':open'),'From "invalid" ("manual") to "auto" should hide the pop-up');
    popUp.showPopUp();
    assert_true(popUp.matches(':open'));
    popUp.setAttribute('popup','invalid');
    assert_false(popUp.matches(':open'),'From "auto" to "invalid" (which is interpreted as "manual") should close the pop-up');
  },'Changing attribute values should close open pop-ups');

  function modalPseudoSupported() {
    try {
      document.createElement('dialog').matches(':modal');
      return true; // No exception means :modal is supported.
    } catch(e) {
      return false;
    }
  }

  ["auto","hint","manual"].forEach(type => {
    test((t) => {
      const popUp = createPopUp(t);
      popUp.setAttribute('popup',type);
      popUp.showPopUp();
      assert_true(popUp.matches(':open'));
      popUp.remove();
      assert_false(popUp.matches(':open'));
      document.body.appendChild(popUp);
      assert_false(popUp.matches(':open'));
    },`Removing a visible popup=${type} element from the document should close the pop-up`);

    if (modalPseudoSupported()) {
      test((t) => {
        const popUp = createPopUp(t);
        popUp.setAttribute('popup',type);
        popUp.showPopUp();
        assert_true(popUp.matches(':open'));
        assert_false(popUp.matches(':modal'));
        popUp.hidePopUp();
      },`A showing popup=${type} does not match :modal`);
    }
  });

  promise_test(async () => {
    const popUp = document.querySelector('[popup].animated');
    assert_true(!!popUp);
    assert_false(isElementVisible(popUp));
    popUp.showPopUp();
    assert_true(popUp.matches(':open'));
    assert_true(getComputedStyle(popUp).opacity < 0.1,'Animations should start on show');
    assert_throws_dom("InvalidStateError",() => popUp.showPopUp(),'Calling showPopUp on a popup that is in the process of animating show should throw InvalidStateError');
    await finishAnimations(popUp);
    assert_true(getComputedStyle(popUp).opacity > 0.9);
    assert_true(isElementVisible(popUp));
    popUp.hidePopUp();
    assert_false(popUp.matches(':open'));
    assert_true(getComputedStyle(popUp).opacity > 0.9,'Animations should start on hide');
    assert_throws_dom("InvalidStateError",() => popUp.hidePopUp(),'Calling hidePopUp on a popup that is in the process of animating hide should throw InvalidStateError');
    popUp.showPopUp(); // But showPopUp should still be ok.
    popUp.hidePopUp(); // Clean up
    await finishAnimations(popUp);
  },'Exceptions are thrown even when show/hide are animated')
</script>
